Statefull/requirements.txt
Brotli
GitPython
Jinja2
Markdown
MarkupSafe
PyJWT
PySocks
PyYAML
Pygments
SQLAlchemy
Werkzeug
absl-py
agent
agents
aiohappyeyeballs
aiohttp
aiosignal
aiosqlite
altair
annotated-types
anthropic
anyio
asttokens
astunparse
async-timeout
attrs
audio-recorder-streamlit
audioread
backcall
backoff
bcrypt
beautifulsoup4
bleach
blinker
blockbuster
build
cachetools
certifi
cffi
charset-normalizer
chromadb
click
cloudpickle
colorama
coloredlogs
comm
contourpy
cryptography
cycler
dataclasses
dataclasses-json
datasets
debugpy
decorator
defusedxml
deep-translator
dill
distro
docopt
duckduckgo_search
durationpy
entrypoints
exceptiongroup
executing
fastjsonschema
fastapi
ffmpeg-python
filelock
filetype
flatbuffers
fonttools
forbiddenfruit
frozenlist
fsspec
future
gTTS
gast
gitdb
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-genai
google-generativeai
google-pasta
googleapis-common-protos
googletrans
grandalf
greenlet
grpcio
grpcio-status
gym
gym-notices
h11
h2
h5py
hpack
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface_hub
humanfriendly
hyperframe
idna
importlib_metadata
ipykernel
ipython
ipython_pygments_lexers
imageio-ffmpeg
jedi
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jsonschema_rs
jupyter_client
jupyter_core
jupyterlab_pygments
tf-keras
kiwisolver
kubernetes
langchain
langchain-anthropic
langchain-chroma
langchain-community
langchain-core
langchain-experimental
langchain-google-genai
langchain-huggingface
langchain-openai
langchain-text-splitters
langchain-groq
langdetect
langfuse
langgraph
langgraph-api
langgraph-checkpoint
langgraph-checkpoint-sqlite
langgraph-cli
langgraph-prebuilt
langgraph-runtime-inmem
langgraph-sdk
langsmith
lazy_loader
libclang
librosa
llvmlite
lxml
markdown-it-py
marshmallow
matplotlib
matplotlib-inline
mdurl
mistune
ml_dtypes
mmh3
moviepy
mpmath
msgpack
multidict
multiprocess
munkres
mypy_extensions
namex
narwhals
nbclient
nbconvert
nbformat
nest_asyncio
networkx
numba
numpy
oauthlib
onnx-asr
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opt_einsum
optree
orjson
ormsgpack
overrides
openai-whisper
packaging
pandas
pandocfilters
parso
pedalboard
pickleshare
pillow
pipreqs
platformdirs
pooch
posthog
primp
proglog
prompt_toolkit
propcache
proto-plus
protobuf
psutil
pure_eval
pyOpenSSL
pyarrow
pyasn1
pyasn1_modules
pybase64
pybind11
pybind11_global
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
pydub
pyparsing
pypdf
pypika
pyproject_hooks
pyreadline3
python-dateutil
python-dotenv
pytz
pyu2f
pyzmq
referencing
regex
requests
requests-oauthlib
requests-toolbelt
responses
rich
rpds-py
rsa
ruamel.yaml
ruamel.yaml.clib
safetensors
scikit-learn
scipy
sentence-transformers
shellingham
six
smmap
sniffio
soundfile
soupsieve
soxr
sqlite-vec
sse-starlette
stack_data
standard-aifc
standard-sunau
starlette
streamlit
streamlit-audiorec
streamlit-audiorecorder
streamlit-autorefresh
structlog
sympy
tenacity
tensorboard
tensorboard-data-server
tensorflow
termcolor
threadpoolctl
tiktoken
tinycss2
tokenizers
toml
toolz
tornado
torch
tqdm
traitlets
transformers
truststore
typer
typing-inspection
typing_extensions
typing_inspect
tzdata
unicodedata2
uritemplate
urllib3
uvicorn
uvicorn[standard]
watchdog == 4.0.0
watchfiles
wcwidth
webencodings
websocket-client
websockets
win_inet_pton
wrapt
xxhash
yarg
yarl
zipp
zstandard

pysqlite3-binary>=0.0.6
onnx-asr[cpu,hub]

# Install the current project in editable mode (equivalent to pip install -e .)
-e .

